requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
PyJWT
Werkzeug
python-dotenv